backend/requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
boto3==1.35.76
botocore==1.35.76
cachetools==4.2.4
certifi==2024.8.30
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.2
django-cors-headers==4.5.0
django-filter==24.3
django-googledrive-storage==1.6.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-api-core==2.10.2
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
httptools==0.6.4
idna==3.10
jmespath==1.0.1
msgpack==1.1.0
msgpack-python==0.5.6
packaging==24.2
protobuf==4.25.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
rsa==4.9
s3transfer==0.10.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.2.3
user-agents==2.2.0
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
